# src/main/resources/swagger.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Flight & Airport API",
    "description": "Flight and Airport CRUD and search API.",
    "contact": {
      "name": "Cemre Şenyuva",
      "url": "https://cemrecodes.github.io",
      "email": "cemressenyuva@gmail.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/flights": {
      "get": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Get all flights",
        "description": "get-flight",
        "operationId": "getCollectionResource-flight-get_1_1",
        "parameters": [
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelFlight"
                }
              },
              "application/x-spring-data-compact+json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelEntityModelFlight"
                }
              },
              "text/uri-list": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Save flight",
        "description": "create-flight",
        "operationId": "postCollectionResource-flight-post_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Flight is successfully created.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelFlight"
                }
              }
            }
          },
          "400": {
            "description": "Departure Airport/Arrival Airport do not exist.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "put": {
        "tags": [
          "User Controller"
        ],
        "summary": "Update user",
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User has been successfully updated.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User Controller"
        ],
        "summary": "User egister",
        "description": "create-user",
        "operationId": "postCollectionResource-user-post_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User has been successfully registered.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelUser"
                }
              }
            }
          },
          "405": {
            "description": "Username taken",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User Controller"
        ],
        "summary": "Delete user",
        "operationId": "deleteUser",
        "responses": {
          "204": {
            "description": "User has been successfully deleted."
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/flights/{flightId}": {
      "get": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Get flight by ID",
        "operationId": "getFlight",
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flight exists and is returned.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Flight doesn't exist",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Update flight by ID",
        "operationId": "updateFlight",
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Flight has been updated successfully.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Flight/Departure Airport/Arrival Airport doesn't exist.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Delete flight by ID",
        "operationId": "deleteFlight",
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Flight has been successfully deleted."
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/users/login": {
      "post": {
        "tags": [
          "User Controller"
        ],
        "summary": "User login",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logged in successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/flights/search": {
      "get": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Search flight",
        "operationId": "searchFlight",
        "parameters": [
          {
            "name": "departureAirport",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arrivalAirport",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departureDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "arrivalDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flight exists and is returned.",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Flight"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Flight"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/flights/addFlightsWithFaker": {
      "get": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Add mock flights",
        "operationId": "addFlightsWithFaker",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Flight"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/flights/addFlightWithFaker": {
      "get": {
        "tags": [
          "Flight Controller"
        ],
        "summary": "Add mock flight",
        "operationId": "addFlightWithFaker",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
...
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "EntityModelUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonLocked": {
            "type": "boolean"
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Airport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string"
          }
        }
      },
...
      "Flight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "departureAirport": {
            "$ref": "#/components/schemas/Airport"
          },
          "arrivalAirport": {
            "$ref": "#/components/schemas/Airport"
          },
          "departureDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "EntityModelFlight": {
        "type": "object",
        "properties": {
          "departureDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "PagedModelEntityModelFlight": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "flights": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelFlight"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "FlightDto": {
        "type": "object",
        "properties": {
          "departureAirport": {
            "type": "string"
          },
          "arrivalAirport": {
            "type": "string"
          },
          "departureDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}